abstract class to share NotificationOpenedReceiver
Created abstract NotificationOpenedReceiverBase to share code between
both activities. It is very unlikly the code would ever be different
between them and prevents possible future mistakes of editing only one.
