+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/NotificationOpenedReceiverBase.kt
@@ -26,14 +26,20 @@
  */
 package com.onesignal
 
-/**
- * This is the same as NotificationOpenedReceiver expect it doesn't contain
- * android:taskAffinity="" in it's AndroidManifest.xml entry. This is to
- * account for the resume behavior not working on Android API 22 and older.
- *
- * In Android 5.x and older, android:taskAffinity="" starts a new task as
- * expected, however the allowTaskReparenting behavior does not happen which
- * results in the app not resuming, when using reverse Activity trampoline.
- * Oddly enough cold starts of the app were not a problem.
- */
-class NotificationOpenedReceiverAndroid22AndOlder : NotificationOpenedReceiverBase()
+import android.app.Activity
+import android.content.Intent
+import android.os.Bundle
+
+abstract class NotificationOpenedReceiverBase : Activity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        NotificationOpenedProcessor.processFromContext(this, intent)
+        finish()
+    }
+
+    override fun onNewIntent(intent: Intent) {
+        super.onNewIntent(intent)
+        NotificationOpenedProcessor.processFromContext(this, getIntent())
+        finish()
+    }
+}
